Make the mock_device usable as an executable
This executable can be used with socat for creating a mock serial device
on a Linux system.

There are only a handful of basic request/reply strings available in the
mock device. It is not a complete representation of the way a RAVEn
device behaves, but is still useful for smoke testing.