Defer feeding XML data until we likely trigger an event (#29)
Recent changes to expat, the XML parser behind the Python XMLPullParser,
utilize a backoff when feeding the parser in small chunks which don't
(yet) constitute something interesting. Unfortunately, this could result
in the parser sitting on a completed payload waiting for more data
before it attempts parsing.

The safest mitigation I can find is to just wait until we get a '>'
before feeding the parser, which should hopefully mean that we'll never
trigger the backoff.