Handle filler value for DateCode (#38)
It appears that some EMU-2 devices will fill the date code with '\xff'
characters in some scenarios. We've already seen some fields use 0xff to
represent a sort of 'filler' value where there is no date to give. This
scenario is a little bit different because the raw representation is
0xff where before we were seeing 0xff written out in ASCII characters,
but given that the number of characters still matches documentation's
expected value I think we can draw the same conclusion here.

There are a few components to this change:
1. Switch from plain ASCII to codepage 1252 extended ASCII. This is
   actually called out in the RAVEn protocol documentation.
2. Set the _stash attribute earlier in the _reset method to limit
   the fallout of any problems feeding the XML root tag to the parser.
3. Specifically handle an all-0xFF date code as 'None'.